updated query tool
