--- src/components/data-viz/DataTable/index.js
@@ -1,1 +1,0 @@
-export default from './DataTable'
