+++ src/components/data-viz/QueryToolTable/Custom/AllTypeProvider.js
