+++ src/components/data-viz/QueryToolTable/index.js
@@ -1,0 +1,1 @@
+export default from './QueryToolTable'
