+++ src/js/query-manager/query-tool-queries.js
@@ -65,10 +65,7 @@ const VARIABLE_CONFIGS = {
   ],
   [PRODUCTION]: [
     { [LAND_TYPE]: MULTI_STR },
-    { [OFFSHORE_REGION]: MULTI_STR },
-    { [US_STATE]: MULTI_STR },
     { [COUNTY]: MULTI_STR },
-    { [COMMODITY]: MULTI_STR },
     { [PRODUCT]: MULTI_STR },
     { [STATE_OFFSHORE_NAME]: MULTI_STR },
     { [PERIOD]: SINGLE_STR },
@@ -119,11 +116,11 @@ export const getVariables = (state, options) => getVariableValues(state)
 // STEP 2: Define all the queries needed
 
 const VIEWS = {
-  [REVENUE]: 'query_tool_revenue_data',
-  [PRODUCTION]: 'query_tool_production_data',
-  [DISBURSEMENT]: 'query_tool_disbursement_data',
+  [REVENUE]: 'query_tool_revenue',
+  [PRODUCTION]: 'query_tool_production',
+  [DISBURSEMENT]: 'query_tool_disbursement',
   [DISBURSEMENT_MONTHLY]: 'query_tool_disbursement_monthly_data',
-  [REVENUE_BY_COMPANY]: 'query_tool_revenue_by_company'
+  [REVENUE_BY_COMPANY]: 'query_tool_fed_revenue_by_company'
 }
 const REVENUE_QUERY = whereClause => (
   `results:${ VIEWS[REVENUE] }(
@@ -135,7 +132,9 @@ const REVENUE_QUERY = whereClause => (
       ${ LAND_TYPE }: ${ DB_COLS[LAND_TYPE] }
       ${ STATE_OFFSHORE_NAME }: ${ DB_COLS[STATE_OFFSHORE_NAME] }
       ${ COUNTY }: ${ DB_COLS[COUNTY_NAME] }
-      ${ ALL_REVENUE_YEARS }
+      ${ CALENDAR_YEAR }: ${ DB_COLS[CALENDAR_YEAR] }
+      ${ FISCAL_YEAR }: ${ DB_COLS[FISCAL_YEAR] }
+      ${ REVENUE }: ${ DB_COLS[REVENUE] }
     }
     counts:${ VIEWS[REVENUE] }_aggregate (
       where: {
@@ -158,12 +157,13 @@ const PRODUCTION_QUERY = whereClause => (
       ${ whereClause }
     }) {
     ${ PRODUCT }: ${ DB_COLS[PRODUCT] }
-    ${ COMMODITY_ORDER }: ${ DB_COLS[COMMODITY_ORDER] }
     ${ LAND_TYPE }: ${ DB_COLS[LAND_TYPE] }
     ${ STATE_OFFSHORE_NAME }: ${ DB_COLS[STATE_OFFSHORE_NAME] }
     ${ COUNTY }: ${ DB_COLS[COUNTY_NAME] }
     ${ MONTH_LONG }: ${ DB_COLS[MONTH_LONG] }
-    ${ ALL_PRODUCTION_YEARS }
+    ${ CALENDAR_YEAR }: ${ DB_COLS[CALENDAR_YEAR] }
+    ${ FISCAL_YEAR }: ${ DB_COLS[FISCAL_YEAR] }
+    ${ PRODUCTION }: ${ DB_COLS[PRODUCTION] }
   }
   counts:${ VIEWS[PRODUCTION] }_aggregate (
     where: {
@@ -175,36 +175,8 @@ const PRODUCTION_QUERY = whereClause => (
       ${ COUNTY }:count(columns: ${ DB_COLS[COUNTY_NAME] }, distinct: true)
       ${ LAND_TYPE }:count(columns: ${ DB_COLS[LAND_TYPE] }, distinct: true)
       ${ STATE_OFFSHORE_NAME }:count(columns: ${ DB_COLS[STATE_OFFSHORE_NAME] }, distinct: true)
-      ${ MONTH_LONG }:count(columns: ${ DB_COLS[MONTH_LONG] }, distinct: true)
     }
   }`)
-
-const PRODUCTION_MONTHLY_QUERY = whereClause => (
-  `results:${ VIEWS[PRODUCTION] }(
-      where: {
-        ${ whereClause }
-      }) {
-      ${ PRODUCT }: ${ DB_COLS[PRODUCT] }
-      ${ COMMODITY_ORDER }: ${ DB_COLS[COMMODITY_ORDER] }
-      ${ LAND_TYPE }: ${ DB_COLS[LAND_TYPE] }
-      ${ STATE_OFFSHORE_NAME }: ${ DB_COLS[STATE_OFFSHORE_NAME] }
-      ${ COUNTY }: ${ DB_COLS[COUNTY_NAME] }
-      ${ MONTH_LONG }: ${ DB_COLS[MONTH_LONG] }
-      ${ ALL_PRODUCTION_MONTHLY_YEARS }
-    }
-    counts:${ VIEWS[PRODUCTION] }_aggregate (
-      where: {
-        ${ whereClause }
-      }
-      ) {
-      aggregate {
-        ${ PRODUCT }:count(columns: ${ DB_COLS[PRODUCT] }, distinct: true)
-        ${ COUNTY }:count(columns: ${ DB_COLS[COUNTY_NAME] }, distinct: true)
-        ${ LAND_TYPE }:count(columns: ${ DB_COLS[LAND_TYPE] }, distinct: true)
-        ${ STATE_OFFSHORE_NAME }:count(columns: ${ DB_COLS[STATE_OFFSHORE_NAME] }, distinct: true)
-        ${ MONTH_LONG }:count(columns: ${ DB_COLS[MONTH_LONG] }, distinct: true)
-      }
-    }`)
 
 const DISBURSEMENT_QUERY = whereClause => (
   `results:${ VIEWS[DISBURSEMENT] }(
@@ -215,7 +187,10 @@ const DISBURSEMENT_QUERY = whereClause => (
     ${ SOURCE }: ${ DB_COLS[SOURCE] }
     ${ US_STATE }: ${ DB_COLS[US_STATE_NAME] }
     ${ LOCAL_RECIPIENT }: ${ DB_COLS[LOCAL_RECIPIENT] }
-    ${ ALL_DISBURSEMENT_YEARS }
+    ${ MONTH_LONG }: ${ DB_COLS[MONTH_LONG] }
+    ${ CALENDAR_YEAR }: ${ DB_COLS[CALENDAR_YEAR] }
+    ${ FISCAL_YEAR }: ${ DB_COLS[FISCAL_YEAR] }
+    ${ DISBURSEMENT }: ${ DB_COLS[DISBURSEMENT] }
   }
   counts:${ VIEWS[DISBURSEMENT] }_aggregate (
     where: {
@@ -230,32 +205,6 @@ const DISBURSEMENT_QUERY = whereClause => (
     }
   }`)
 
-const DISBURSEMENT_MONTHLY_QUERY = whereClause => (
-  `results:${ VIEWS[DISBURSEMENT_MONTHLY] }(
-      where: {
-        ${ whereClause }
-      }) {
-      ${ RECIPIENT }: ${ DB_COLS[RECIPIENT] }
-      ${ SOURCE }: ${ DB_COLS[SOURCE] }
-      ${ US_STATE }: ${ DB_COLS[US_STATE_NAME] }
-      ${ LOCAL_RECIPIENT }: ${ DB_COLS[LOCAL_RECIPIENT] }
-      ${ MONTH_LONG }: ${ DB_COLS[MONTH_LONG] }
-      ${ ALL_DISBURSEMENT_YEARS }
-    }
-    counts:${ VIEWS[DISBURSEMENT_MONTHLY] }_aggregate (
-      where: {
-        ${ whereClause }
-      }
-      ) {
-      aggregate {
-        ${ RECIPIENT }:count(columns: ${ DB_COLS[RECIPIENT] }, distinct: true)
-        ${ SOURCE }:count(columns: ${ DB_COLS[SOURCE] }, distinct: true)
-        ${ US_STATE }:count(columns: ${ DB_COLS[US_STATE_NAME] }, distinct: true)
-        ${ LOCAL_RECIPIENT }:count(columns: ${ DB_COLS[LOCAL_RECIPIENT] }, distinct: true)
-        ${ MONTH_LONG }:count(columns: ${ DB_COLS[MONTH_LONG] }, distinct: true)
-      }
-    }`)
-
 const REVENUE_BY_COMPANY_QUERY = whereClause => (
   `results:${ VIEWS[REVENUE_BY_COMPANY] }(
     where: {
@@ -264,13 +213,14 @@ const REVENUE_BY_COMPANY_QUERY = whereClause => (
       ${ COMMODITY }: ${ DB_COLS[COMMODITY] }
       ${ REVENUE_TYPE }: ${ DB_COLS[REVENUE_TYPE] }
       ${ COMPANY_NAME }: ${ DB_COLS[COMPANY_NAME] }
-      ${ ALL_REVENUE_BY_COMPANY_YEARS }  }
+      ${ CALENDAR_YEAR }: ${ DB_COLS[CALENDAR_YEAR] }
+      ${ REVENUE }: ${ DB_COLS[REVENUE] }
+    }
   counts:${ VIEWS[REVENUE_BY_COMPANY] }_aggregate (
     where: {
       ${ whereClause }
     }) {
     aggregate {
-      ${ PERIOD }:count(columns: ${ DB_COLS[PERIOD] }, distinct: true)
       ${ COMMODITY }:count(columns: ${ DB_COLS[COMMODITY] }, distinct: true)
       ${ REVENUE_TYPE }:count(columns: ${ DB_COLS[REVENUE_TYPE] }, distinct: true)
       ${ COMPANY_NAME }:count(columns: ${ DB_COLS[COMPANY_NAME] }, distinct: true)
@@ -294,34 +244,14 @@ const QUERIES = {
     gql`query GetDataTableRevenue
       (${ getDataFilterVariableList(state, variableConfig) })
       {${ REVENUE_QUERY(getDataFilterWhereClauses(variableConfig)) }}`,
-  [PRODUCTION]: (state, variableConfig) => {
-    if (state[PERIOD] === PERIOD_MONTHLY) {
-      return (
-        gql`query GetDataTableProductionMonthly
-        (${ getDataFilterVariableList(state, variableConfig) })
-        {${ PRODUCTION_MONTHLY_QUERY(getDataFilterWhereClauses(variableConfig)) }}`
-      )
-    }
-    return (
-      gql`query GetDataTableProduction
+  [PRODUCTION]: (state, variableConfig) =>
+    gql`query GetDataTableProduction
       (${ getDataFilterVariableList(state, variableConfig) })
-      {${ PRODUCTION_QUERY(getDataFilterWhereClauses(variableConfig)) }}`
-    )
-  },
-  [DISBURSEMENT]: (state, variableConfig) => {
-    if (state[PERIOD] === PERIOD_MONTHLY) {
-      return (
-        gql`query GetDataTableDisbursementMonthly
-        (${ getDataFilterVariableList(state, variableConfig) })
-        {${ DISBURSEMENT_MONTHLY_QUERY(getDataFilterWhereClauses(variableConfig)) }}`
-      )
-    }
-    return (
-      gql`query GetDataTableDisbursement
+      {${ PRODUCTION_QUERY(getDataFilterWhereClauses(variableConfig)) }}`,
+  [DISBURSEMENT]: (state, variableConfig) =>
+    gql`query GetDataTableDisbursement
       (${ getDataFilterVariableList(state, variableConfig) })
-      {${ DISBURSEMENT_QUERY(getDataFilterWhereClauses(variableConfig)) }}`
-    )
-  },
+      {${ DISBURSEMENT_QUERY(getDataFilterWhereClauses(variableConfig)) }}`,
   [REVENUE_BY_COMPANY]: (state, variableConfig) =>
     gql`query GetDataTableRevenueByCompany
           (${ getDataFilterVariableList(state, variableConfig) })
