+++ src/pages/query-data.mdx
@@ -11,7 +11,7 @@ includeToc: false
     <QueryTableToolbar label='Query data'/>
   </Grid>
   <Grid item xs={12}>
-    <DataTable />
+    <QueryToolTable />
   </Grid>
 </Grid>
