+++ src/stores/data-filter-store/index.js
@@ -26,7 +26,7 @@ function DataFilterProvider ({ children, defaults, urlParams }) {
   const actions = useActions(state, dispatch)
 
   // Log new state
-  useEffect(() => console.log('DataFilterProvider newState: ', state), [state])
+  // useEffect(() => console.log('DataFilterProvider newState: ', state), [state])
 
   return (
     <DataFilterContext.Provider value={{ state, ...actions }}>
